+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/model/Member.java
@@ -2,9 +2,9 @@ package com.bunnyxt.tdd.model;
 
 public class Member {
 
-    private int id;
-    private int added;
-    private int mid;
+    private Long id;
+    private Integer added;
+    private Integer mid;
     private String sex;
     private String name;
     private String face;
@@ -15,27 +15,27 @@ public class Member {
 
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getAdded() {
+    public Integer getAdded() {
         return added;
     }
 
-    public void setAdded(int added) {
+    public void setAdded(Integer added) {
         this.added = added;
     }
 
-    public int getMid() {
+    public Integer getMid() {
         return mid;
     }
 
-    public void setMid(int mid) {
+    public void setMid(Integer mid) {
         this.mid = mid;
     }
