+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/model/VideoRecord.java
@@ -2,94 +2,94 @@ package com.bunnyxt.tdd.model;
 
 public class VideoRecord {
 
-    private int id;
-    private int added;
-    private int aid;
-    private int view;
-    private int danmaku;
-    private int reply;
-    private int favorite;
-    private int coin;
-    private int share;
-    private int like;
+    private Long id;
+    private Integer added;
+    private Integer aid;
+    private Integer view;
+    private Integer danmaku;
+    private Integer reply;
+    private Integer favorite;
+    private Integer coin;
+    private Integer share;
+    private Integer like;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getAdded() {
+    public Integer getAdded() {
         return added;
     }
 
-    public void setAdded(int added) {
+    public void setAdded(Integer added) {
         this.added = added;
     }
 
-    public int getAid() {
+    public Integer getAid() {
         return aid;
     }
 
-    public void setAid(int aid) {
+    public void setAid(Integer aid) {
         this.aid = aid;
     }
 
-    public int getView() {
+    public Integer getView() {
         return view;
     }
 
-    public void setView(int view) {
+    public void setView(Integer view) {
         this.view = view;
     }
 
-    public int getDanmaku() {
+    public Integer getDanmaku() {
         return danmaku;
     }
 
-    public void setDanmaku(int danmaku) {
+    public void setDanmaku(Integer danmaku) {
         this.danmaku = danmaku;
     }
 
-    public int getReply() {
+    public Integer getReply() {
         return reply;
     }
 
-    public void setReply(int reply) {
+    public void setReply(Integer reply) {
         this.reply = reply;
     }
 
-    public int getFavorite() {
+    public Integer getFavorite() {
         return favorite;
     }
 
-    public void setFavorite(int favorite) {
+    public void setFavorite(Integer favorite) {
         this.favorite = favorite;
     }
 
-    public int getCoin() {
+    public Integer getCoin() {
         return coin;
     }
 
-    public void setCoin(int coin) {
+    public void setCoin(Integer coin) {
         this.coin = coin;
     }
 
-    public int getShare() {
+    public Integer getShare() {
         return share;
     }
 
-    public void setShare(int share) {
+    public void setShare(Integer share) {
         this.share = share;
     }
 
-    public int getLike() {
+    public Integer getLike() {
         return like;
     }
 
-    public void setLike(int like) {
+    public void setLike(Integer like) {
         this.like = like;
     }
